Process incomplete ASCII symbols
Console pump reads output from QEMU in small chunks. It might happen
that ASCII sequence (that might be as long as 20 characters) partially
pumped to the buffer. But ansiRe fails to process this incomplete
sequence. To handle such situation we check what is the start index of
the last unhandled ASCII sequence. If it is close enough to the end of
the pumped buffer then we leave this small unprocessed sequence in the buffer.
The processed sequences are printed to stdout.